+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds `/power :level` command to set your power level for a guild.
+
 ## [v7.3.3](https://github.com/lexicalunit/spellbot/releases/tag/v7.3.3) - 2021-11-18
 
 ### Fixed